Move database values to types
