+++ frontend/src/metabase-types/api/database.ts
@@ -9,11 +9,8 @@ export type DatabaseSettings = {
   [key: string]: any;
 };
 
-export interface Database {
+export interface Database extends DatabaseValues {
   id: DatabaseId;
-  name: string;
-  engine: string;
-  is_sample: boolean;
   is_saved_questions: boolean;
   creator_id?: number;
   created_at: string;
@@ -26,6 +23,20 @@ export interface Database {
 
   // Only appears in  GET /api/database/:id
   "can-manage"?: boolean;
+}
+
+export interface DatabaseValues {
+  id?: DatabaseId;
+  name: string;
+  engine: string | undefined;
+  details: Record<string, unknown>;
+  schedules: DatabaseSchedules;
+  auto_run_queries: boolean | null;
+  refingerprint: boolean | null;
+  cache_ttl: number | null;
+  is_sample: boolean;
+  is_full_sync: boolean;
+  is_on_demand: boolean;
 }
 
 export interface DatabaseSchedules {
