+++ frontend/src/metabase-types/api/timeline.ts
@@ -6,19 +6,13 @@ export type TimelineId = number;
 export type TimelineEventId = number;
 export type TimelineEventSource = "question" | "collections" | "api";
 
-export interface Timeline {
+export interface Timeline extends TimelineValues {
   id: TimelineId;
-  collection_id: RegularCollectionId | null;
-  name: string;
-  description: string | null;
-  icon: string;
-  default: boolean;
-  archived: boolean;
   collection?: Collection;
   events?: TimelineEvent[];
 }
 
-export interface TimelineData {
+export interface TimelineValues {
   id?: TimelineId;
   collection_id: RegularCollectionId | null;
   name: string;
@@ -28,16 +22,9 @@ export interface TimelineData {
   archived: boolean;
 }
 
-export interface TimelineEvent {
+export interface TimelineEvent extends TimelineEventData {
   id: TimelineEventId;
   timeline_id: TimelineId;
-  name: string;
-  description: string | null;
-  icon: string;
-  timestamp: string;
-  timezone: string;
-  time_matters: boolean;
-  archived: boolean;
   creator: User;
   created_at: string;
 }
