--- frontend/src/metabase/admin/databases/components/widgets/EngineWidget/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./EngineWidget";
