--- frontend/src/metabase/admin/databases/components/widgets/EngineWidget/types.ts
@@ -1,10 +1,0 @@
-export interface EngineField {
-  value?: string;
-  onChange?: (value: string | undefined) => void;
-}
-
-export interface EngineOption {
-  name: string;
-  value: string;
-  index: number;
-}
