+++ frontend/src/metabase/databases/components/DatabaseCacheScheduleField/DatabaseCacheScheduleField.tsx
@@ -3,8 +3,11 @@ import { useField, useFormikContext } from "formik";
 import { t } from "ttag";
 import FormField from "metabase/core/components/FormField";
 import SchedulePicker from "metabase/components/SchedulePicker";
-import { ScheduleSettings, ScheduleType } from "metabase-types/api";
-import { DatabaseValues } from "../../types";
+import {
+  DatabaseValues,
+  ScheduleSettings,
+  ScheduleType,
+} from "metabase-types/api";
 import {
   ScheduleOptionList,
   ScheduleOptionBody,
