+++ frontend/src/metabase/databases/components/DatabaseClientIdDescription/DatabaseClientIdDescription.tsx
@@ -2,7 +2,7 @@ import React from "react";
 import { useFormikContext } from "formik";
 import { jt, t } from "ttag";
 import ExternalLink from "metabase/core/components/ExternalLink";
-import { DatabaseValues } from "../../types";
+import { DatabaseValues } from "metabase-types/api";
 
 const CREDENTIAL_URLS: Record<string, string> = {
   bigquery:
