+++ frontend/src/metabase/databases/types.ts
@@ -1,24 +1,5 @@
 import { ComponentType, ReactNode } from "react";
-import {
-  DatabaseId,
-  DatabaseSchedules,
-  EngineFieldOption,
-  EngineFieldType,
-} from "metabase-types/api";
-
-export interface DatabaseValues {
-  id?: DatabaseId;
-  name: string;
-  engine: string | undefined;
-  details: Record<string, unknown>;
-  schedules: DatabaseSchedules;
-  auto_run_queries: boolean | null;
-  refingerprint: boolean | null;
-  cache_ttl: number | null;
-  is_sample: boolean;
-  is_full_sync: boolean;
-  is_on_demand: boolean;
-}
+import { EngineFieldOption, EngineFieldType } from "metabase-types/api";
 
 export interface EngineOption {
   name: string;
