+++ frontend/src/metabase/timelines/common/components/TimelineForm/TimelineForm.tsx
@@ -11,7 +11,7 @@ import FormSelect from "metabase/core/components/FormSelect";
 import FormSubmitButton from "metabase/core/components/FormSubmitButton";
 import FormErrorMessage from "metabase/core/components/FormErrorMessage";
 import * as Errors from "metabase/core/utils/errors";
-import { TimelineData } from "metabase-types/api";
+import { TimelineValues } from "metabase-types/api";
 import FormArchiveButton from "../FormArchiveButton";
 import { TimelineFormFooter } from "./TimelineForm.styled";
 
@@ -22,8 +22,8 @@ const TIMELINE_SCHEMA = Yup.object({
 });
 
 export interface TimelineFormProps {
-  initialValues: TimelineData;
-  onSubmit: (data: TimelineData) => void;
+  initialValues: TimelineValues;
+  onSubmit: (data: TimelineValues) => void;
   onArchive?: () => void;
   onCancel?: () => void;
 }
